Trait governance_os_pallet_bylaws::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type Role: Parameter + Member + MaybeSerializeDeserialize + Ord + Default;
    type WeightInfo: WeightInfo;
    type MaxRoles: Get<u32>;
    type RoleBuilder: RoleBuilder<Role = Self::Role>;
}

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Role: Parameter + Member + MaybeSerializeDeserialize + Ord + Default

Roles defines UNIX like roles that users must be granted before triggering certain calls.

type WeightInfo: WeightInfo

The weights for this pallet.

type MaxRoles: Get<u32>

Only used for weight calculations: this is the highest number of roles we expect one account to have.

type RoleBuilder: RoleBuilder<Role = Self::Role>

Helper for the runtime to specify its custom roles.

Loading content...